LLM my own/PDF chatbot/requirements.txt
langchain
streamlit
huggingface_hub
tiktoken
python-dotenv
faiss-cpu
transformers
PyPDF2
